+++ doc/changes.rst
@@ -5,6 +5,14 @@ desitarget Change Log
 0.47.1 (unreleased)
 -------------------
 
+* First run of secondaries with real target files [`PR #669`_]. Includes:
+    * Add Gaia-only standard stars to the MWS masks for SV, Main Survey:
+        * `GAIA_STD_FAINT`, `GAIA_STD_BRIGHT`, `GAIA_STD_WD`.
+    * General optimization, updating and debugging of the secondary code.
+    * Get `TARGETIDs` from the input sweeps, not just the actual targets.
+    * Add the first full bitmask for the SV1 secondary target files.
+    * Updates to the data model to better reflect the primary targets.
+* Clean-up minor style and doc issues from `PR #636`_ [`PR #668`_].
 * Updates and bug fixes for DR9 now SV is on-sky [`PR #665`_]. Includes:
     * Pass `MASKBITS` column forward for GFAs.
     * Bug fixes necessitated by target files having a second extension.
@@ -15,12 +23,12 @@ desitarget Change Log
 * Simplify stellar SV bits [`PR #636`_]:
     * Secondary bit requirement for main stellar SV program to 4 bits.
     * Primary bright science WDs use the old algorithmic selection.
-* Clean-up minor style and doc issues from `PR #636`_ [`PR #668`_].
 
 .. _`PR #636`: https://github.com/desihub/desitarget/pull/636
 .. _`PR #664`: https://github.com/desihub/desitarget/pull/664
 .. _`PR #665`: https://github.com/desihub/desitarget/pull/665
 .. _`PR #668`: https://github.com/desihub/desitarget/pull/668
+.. _`PR #669`: https://github.com/desihub/desitarget/pull/669
 
 0.47.0 (2020-12-10)
 -------------------